config/mersey-green/query.rq - don't explode!
This query has been exploding: generatign a huge result of thousands
of rows, when there are only about 200 actual initiatives!

The cause is cominatorial: when there is no geocoded location, the Geo
Container field is absent, and this causes a match in the SPARQL query
from each initiative to all of the other initiatives' locations.

The solution is the same as in the ICA map: only join new data on join
fields which have been successfully joined. Which means: in the SPARQL
query, put the latter in blocks, preceeded by the former.